# server/openapi.yaml
openapi: 3.0.0
info:
  title: Mystická Hvězda API
  description: API pro ezoterické služby, horoskopy a tarotové výklady.
  version: 1.0.0
servers:
  - url: http://localhost:3001/api
    description: Lokální vývojový server
  - url: https://www.mystickahvezda.cz/api
    description: Produkční server
paths:
...
  /analytics/event:
    post:
      summary: Record first-party analytics event
      description: Records a whitelisted browser or smoke-test event. Requires CSRF token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalyticsEventRequest'
      responses:
        '200':
          description: Event accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsAck'
        '400':
          description: Invalid event name or payload.
        '403':
          description: CSRF token missing or invalid.
  /analytics/batch:
    post:
      summary: Record first-party analytics batch
      description: Records up to 10 whitelisted analytics events in one CSRF-protected request.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalyticsBatchRequest'
      responses:
        '200':
          description: Batch accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsAck'
        '400':
          description: No valid analytics events.
        '403':
          description: CSRF token missing or invalid.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AnalyticsEventRequest:
      type: object
      required: [eventName]
      properties:
        eventName:
          type: string
          example: cta_clicked
        name:
          type: string
          description: Legacy alias for eventName.
        feature:
          type: string
          nullable: true
        page:
          type: string
          nullable: true
        path:
          type: string
          nullable: true
        referrer:
          type: string
          nullable: true
        clientId:
          type: string
          nullable: true
        sessionId:
          type: string
          nullable: true
        metadata:
          type: object
          additionalProperties: true
    AnalyticsBatchRequest:
      type: object
      required: [events]
      properties:
        events:
          type: array
          maxItems: 10
          items:
            $ref: '#/components/schemas/AnalyticsEventRequest'
    AnalyticsAck:
      type: object
      properties:
        success:
          type: boolean
        accepted:
          type: integer
...